Add a warning when the shape palette is exceeded
This warning was present in ggplot before and is necessary to understand
why shapes beyond the 6th are suppresses. But this being outside of the
scope of ggplot, I did not know how to phrase the warning exactly.

For information, the warning in ggplot was
scale_shape_discrete can deal with a maximum of 6 discrete
values, but you have 11.  See ?scale_manual for a possible alternative