Support units for `breaks_width(offset = )` #247 (#269)
Based on `fullseq()` in  `R/full-seq.r`.

`base::seq()` is used for consistency with the `width` argument.  To add one
unit, a sequence of `length.out = 2, by = unit` is created, and the second
element is taken.  This process is vectorised using `lapply()` and `do.call(c)`,
because the alternatives `vapply()` or `unlist()` strip the attributes of
`Date` and `POSIXt` objects.

This is modified to apply `offset_by()` to the output of `fullseq()`.  To
support offstes with compound units, for example one month and five days for the
UK tax year, the `offset` argument can be a vector, and the offsets are applied
sequentially in the order given.  An example illustrates the difference that the
order can make.